dev-demo: correct cers column type
AdminForth/1821/agent-plugins-doesn't-work-out

+++ dev-demo/resources/cars_resources/carsResourseTemplate.ts
@@ -134,7 +134,7 @@ export default function carsResourseTemplate(resourceId: string, dataSource: Car
       {
         name: 'mileage',
         allowMinMaxQuery: true,
-        type: AdminForthDataTypes.DECIMAL,
+        type: AdminForthDataTypes.FLOAT,
       },
       {
         name: 'color',
